+++ tests/functional/Extension/FrontMatterExtension/FrontMatterExtensionTest.php
@@ -157,6 +157,34 @@ EOT;
     /**
      * @dataProvider parserProvider
      */
+    public function testWithWindowsLineEndings(FrontMatterDataParserInterface $parser, bool $shouldTest): void
+    {
+        $this->skipIfParserNotAvailable($parser, $shouldTest);
+
+        $markdown = "---\r\nfoo: bar\r\n---\r\n\r\n# Test";
+
+        $expectedHtml        = "<h1>Test</h1>\n";
+        $expectedFrontMatter = ['foo' => 'bar'];
+
+        $converter = new MarkdownConverter($this->getEnvironment($parser));
+        $result    = $converter->convertToHtml($markdown);
+
+        $this->assertInstanceOf(RenderedContentWithFrontMatter::class, $result);
+        $this->assertInstanceOf(\Stringable::class, $result);
+
+        \assert($result instanceof RenderedContentWithFrontMatter);
+        $this->assertSame($expectedFrontMatter, $result->getFrontMatter());
+
+        $this->assertSame($expectedHtml, (string) $result->getContent());
+        $this->assertSame($expectedHtml, (string) $result);
+
+        $this->assertSame(1, $result->getDocument()->getStartLine());
+        $this->assertSame(5, $result->getDocument()->firstChild()->getStartLine());
+    }
+
+    /**
+     * @dataProvider parserProvider
+     */
     public function testWithNoFrontMatter(FrontMatterDataParserInterface $parser, bool $shouldTest): void
     {
         $this->skipIfParserNotAvailable($parser, $shouldTest);